requirements.txt
fastapi>=0.96.0,<0.97.0
uvicorn[standard]>=0.20.0,<0.21.0
reflex==0.4.3
supabase==2.3.0
pydantic>=1.10.2,<2.0.0
python-dotenv==1.0.0
python-multipart>=0.0.5,<0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT>=2.8.0,<3.0.0
email-validator>=2.0.0,<3.0.0
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0